app: improve run.sh script

+++ run.sh
@@ -1,9 +1,11 @@
 #!/bin/bash
 
 # running postgres
-mkdir -p $HOME/docker/volumes/postgres
-docker pull postgres:12.1
-docker run --rm --name psql-container -e POSTGRES_PASSWORD=pass123 -p 5432:5432 -v $HOME/docker/volumes/postgres:/var/lib/postgresql/data -d postgres:12.1
+if [ $1 = "start_docker" ]; then
+    mkdir -p $HOME/docker/volumes/postgres
+    docker pull postgres:12.1
+    docker run --rm --name psql-container -e POSTGRES_PASSWORD=pass123 -p 5432:5432 -v $HOME/docker/volumes/postgres:/var/lib/postgresql/data -d postgres:12.1
+fi
 
 # running app
 docker build -t vivaorganica_1 -f Dockerfile .
